Change type of time/timestamps to f64

+++ src/request.rs
@@ -1,7 +1,7 @@
 #[derive(Debug)]
 pub struct Request {
     pub id: usize,
-    pub arrival_time: usize,
-    pub total_service: usize,
-    pub remaining_service: usize,
+    pub arrival_time: f64,
+    pub total_service: f64,
+    pub remaining_service: f64,
 }
